Replaced the shoddy setInterval auto refresh with a proper requestAnimationFrame call.
This will request that the animation function be called before the browser performs the next repaint. The number of callbacks is usually 60 times per second, but will generally match the display refresh rate in most web browsers as per W3C recommendation. requestAnimationFrame() calls are paused in most browsers when running in background tabs or hidden <iframe>s in order to improve performance and battery life.

Whereas the old method brutally ran (regardless of being in focus or not) the JavaScript update call every single second.